+++ src/Exceptionist.php
@@ -16,13 +16,12 @@ declare(strict_types=1);
 namespace Tools;
 
 use ArgumentCountError;
-use BadMethodCallException;
 use ErrorException;
 use Exception;
 use ReflectionFunction;
-use Throwable;
 use Tools\Exception\FileNotExistsException;
 use Tools\Exception\KeyNotExistsException;
+use Tools\Exception\MethodNotExistsException;
 use Tools\Exception\NotInArrayException;
 use Tools\Exception\NotReadableException;
 use Tools\Exception\NotWritableException;
@@ -31,24 +30,25 @@ use Tools\Exception\PropertyNotExistsException;
 use TypeError;
 
 /**
- * Exceptionist.
- * @method static string classExists(string $className, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed fileNotExists(string $filename, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static array isArray($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isBool($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isCallable($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static string isDir(string $filename, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isFloat($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isInt($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isIterable($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isNotArray($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isNotPositive($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isNull($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isObject($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isPositive($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isResource($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isString($value, string $message = '', \Throwable|string $exception = \Exception::class)
- * @method static mixed isUrl($value, string $message = '', \Throwable|string $exception = \Exception::class)
+ * The `Exceptionist`.
+ * @template E of \Exception
+ * @method static string classExists(string $className, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static string fileNotExists(string $filename, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static array isArray($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static bool isBool($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static callable isCallable($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static string isDir(string $filename, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static float isFloat($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static int isInt($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static iterable isIterable($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static mixed isNotArray($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static mixed isNotPositive($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static null isNull($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static object isObject($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static int isPositive($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static resource isResource($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static string isString($value, string $message = '', \Exception|string $exception = \Exception::class)
+ * @method static string isUrl(string $value, string $message = '', \Exception|string $exception = \Exception::class)
  * @since 1.4.1
  */
 class Exceptionist
@@ -90,7 +90,7 @@ class Exceptionist
         }
 
         //Calls the PHP function and gets the result
-        [$arguments, $message, $exception] = $arguments + [[], '', Exception::class];
+        [$arguments, $message, $exception] = $arguments + [[], '', ErrorException::class];
         try {
             if (!is_callable($phpName)) {
                 throw new Exception(sprintf('Function `%s()` does not exist', $phpName));
@@ -116,16 +116,19 @@ class Exceptionist
      * Checks whether an array key exists.
      *
      * If you pass an array of keys, they will all be checked.
-     * @param string|int|array<string|int> $key Key to check or an array of keys
+     * @template Keys of array-key|array-key[]
+     * @param Keys $key Key to check or an array of keys
      * @param array $array An array with keys to check
-     * @param string|null $message The failure message that will be appended to
-     *  the generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return string|int|array<string|int>
-     * @throws \Tools\Exception\KeyNotExistsException|\Throwable
+     * @param string $message The failure message that will be appended to the generated message
+     * @param E|class-string<E> $exception The exception class you want to set
+     * @return Keys
+     * @throws \Tools\Exception\KeyNotExistsException|\Exception
      */
-    public static function arrayKeyExists($key, array $array, ?string $message = '', $exception = KeyNotExistsException::class)
+    public static function arrayKeyExists($key, array $array, string $message = '', $exception = KeyNotExistsException::class)
     {
+        if (func_num_args() > 3) {
+            deprecationWarning('The `$exception` parameter is deprecated and will be removed in a later release');
+        }
         foreach ((array)$key as $name) {
             self::isTrue(array_key_exists($name, $array), $message ?: 'Key `' . $name . '` does not exist', $exception);
         }
@@ -135,15 +138,18 @@ class Exceptionist
 
     /**
      * Checks whether a file or directory exists
-     * @param string $filename Path to the file or directory
-     * @param string|null $message The failure message that will be appended to
-     *  the generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return string
-     * @throws \Tools\Exception\FileNotExistsException|\Throwable
+     * @template ExistingFilename as string
+     * @param ExistingFilename $filename Path to the file or directory
+     * @param string $message The failure message that will be appended to the generated message
+     * @param E|class-string<E> $exception The exception class you want to set
+     * @return ExistingFilename
+     * @throws \Tools\Exception\FileNotExistsException|\Exception
      */
-    public static function fileExists(string $filename, ?string $message = '', $exception = FileNotExistsException::class): string
+    public static function fileExists(string $filename, string $message = '', $exception = FileNotExistsException::class): string
     {
+        if (func_num_args() > 2) {
+            deprecationWarning('The `$exception` parameter is deprecated and will be removed in a later release');
+        }
         self::isTrue(file_exists($filename), $message ?: 'File or directory `' . Filesystem::instance()->rtr($filename) . '` does not exist', $exception);
 
         return $filename;
@@ -151,17 +157,20 @@ class Exceptionist
 
     /**
      * Checks if a value exists in an array
-     * @param mixed $needle The searched value
+     * @template Needle
+     * @param Needle $needle The searched value
      * @param array $haystack The array
-     * @param string|null $message The failure message that will be appended to
-     *  the generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return mixed
-     * @throws \Tools\Exception\NotInArrayException|\Throwable
+     * @param string $message The failure message that will be appended to the generated message
+     * @param E|class-string<E> $exception The exception class you want to set
+     * @return Needle
+     * @throws \Tools\Exception\NotInArrayException|\Exception
      * @since 1.5.8
      */
-    public static function inArray($needle, array $haystack, ?string $message = '', $exception = NotInArrayException::class)
+    public static function inArray($needle, array $haystack, string $message = '', $exception = NotInArrayException::class)
     {
+        if (func_num_args() > 3) {
+            deprecationWarning('The `$exception` parameter is deprecated and will be removed in a later release');
+        }
         $message = $message ?: 'The value' . (is_stringable($needle) ? ' `' . $needle . '`' : '') . ' does not exist in array' . (is_stringable($haystack) ? ' `' . array_to_string($haystack) . '`' : '');
         self::isTrue(in_array($needle, $haystack, true), $message, $exception);
 
@@ -170,15 +179,15 @@ class Exceptionist
 
     /**
      * Checks whether a value is `false`
-     * @param mixed $value The value you want to check
-     * @param string|null $message The failure message that will be appended to the
-     *  generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return mixed
-     * @throws \Throwable
+     * @template FalseValue
+     * @param FalseValue $value The value you want to check
+     * @param string $message The failure message that will be appended to the generated message
+     * @param E|class-string<E> $exception The exception class you want to set
+     * @return FalseValue
+     * @throws \Exception
      * @since 1.5.10
      */
-    public static function isFalse($value, ?string $message = '', $exception = ErrorException::class)
+    public static function isFalse($value, string $message = '', $exception = ErrorException::class)
     {
         self::isTrue(!$value, $message, $exception);
 
@@ -187,17 +196,20 @@ class Exceptionist
 
     /**
      * Checks whether an object is an instance of `$class`
-     * @param object $object The object you want to check
+     * @template InstancedObject as object
+     * @param InstancedObject $object The object you want to check
      * @param string $class The class that the object should be an instance of
-     * @param string|null $message The failure message that will be appended to
-     *  the generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return object
-     * @throws \Tools\Exception\ObjectWrongInstanceException|\Throwable
+     * @param string $message The failure message that will be appended to the generated message
+     * @param E|class-string<E> $exception The exception class you want to set
+     * @return InstancedObject
+     * @throws \Tools\Exception\ObjectWrongInstanceException|\Exception
      * @since 1.4.7
      */
-    public static function isInstanceOf(object $object, string $class, ?string $message = '', $exception = ObjectWrongInstanceException::class): object
+    public static function isInstanceOf(object $object, string $class, string $message = '', $exception = ObjectWrongInstanceException::class): object
     {
+        if (func_num_args() > 3) {
+            deprecationWarning('The `$exception` parameter is deprecated and will be removed in a later release');
+        }
         self::isTrue($object instanceof $class, $message ?: sprintf('Object `%s` is not an instance of `%s`', get_class($object), $class), $exception);
 
         return $object;
@@ -205,17 +217,20 @@ class Exceptionist
 
     /**
      * Checks whether a file or directory exists and is readable
-     * @param string $filename Path to the file or directory
-     * @param string|null $message The failure message that will be appended to
-     *  the generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return string
+     * @template ReadableFilename as string
+     * @param ReadableFilename $filename Path to the file or directory
+     * @param string $message The failure message that will be appended to the generated message
+     * @param E|class-string<E> $exception The exception class you want to set
+     * @return ReadableFilename
      * @throws \Tools\Exception\FileNotExistsException
      * @throws \Tools\Exception\NotReadableException
-     * @throws \Throwable
+     * @throws \Exception
      */
-    public static function isReadable(string $filename, ?string $message = '', $exception = NotReadableException::class): string
+    public static function isReadable(string $filename, string $message = '', $exception = NotReadableException::class): string
     {
+        if (func_num_args() > 2) {
+            deprecationWarning('The `$exception` parameter is deprecated and will be removed in a later release');
+        }
         self::fileExists($filename, $message, $exception);
         self::isTrue(is_readable($filename), $message ?: sprintf('File or directory `%s` is not readable', Filesystem::instance()->rtr($filename)), $exception);
 
@@ -224,17 +239,20 @@ class Exceptionist
 
     /**
      * Checks whether a file or directory exists and is writable
-     * @param string $filename Path to the file or directory
-     * @param string|null $message The failure message that will be appended to
-     *  the generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return string
+     * @template WritableFilename as string
+     * @param WritableFilename $filename Path to the file or directory
+     * @param string $message The failure message that will be appended to the generated message
+     * @param E|class-string<E> $exception The exception class you want to set
+     * @return WritableFilename
      * @throws \Tools\Exception\FileNotExistsException
      * @throws \Tools\Exception\NotWritableException
-     * @throws \Throwable
+     * @throws \Exception
      */
-    public static function isWritable(string $filename, ?string $message = '', $exception = NotWritableException::class): string
+    public static function isWritable(string $filename, string $message = '', $exception = NotWritableException::class): string
     {
+        if (func_num_args() > 2) {
+            deprecationWarning('The `$exception` parameter is deprecated and will be removed in a later release');
+        }
         self::fileExists($filename, $message, $exception);
         self::isTrue(is_writable($filename), $message ?: sprintf('File or directory `%s` is not writable', Filesystem::instance()->rtr($filename)), $exception);
 
@@ -243,17 +261,21 @@ class Exceptionist
 
     /**
      * Checks whether a class method exists
-     * @param class-string|object $object An object instance or a class name
-     * @param string $methodName The method name
-     * @param string|null $message The failure message that will be appended to
-     *  the generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return array<string> Array with class name and method name
-     * @throws \BadMethodCallException|\Throwable
+     * @template ExistingMethod of string
+     * @template RelativeObject as object
+     * @param class-string<RelativeObject>|RelativeObject $object An object instance or a class name
+     * @param ExistingMethod $methodName The method name
+     * @param string $message The failure message that will be appended to the generated message
+     * @param E|class-string<E> $exception The exception class you want to set
+     * @return array{class-string<RelativeObject>, ExistingMethod} Array with class name and method name
+     * @throws \Tools\Exception\MethodNotExistsException|\Exception
      * @since 1.4.3
      */
-    public static function methodExists($object, string $methodName, ?string $message = '', $exception = BadMethodCallException::class): array
+    public static function methodExists($object, string $methodName, string $message = '', $exception = MethodNotExistsException::class): array
     {
+        if (func_num_args() > 3) {
+            deprecationWarning('The `$exception` parameter is deprecated and will be removed in a later release');
+        }
         $object = is_string($object) ? $object : get_class($object);
         self::isTrue(method_exists($object, $methodName), $message ?: sprintf('Method `%s::%s()` does not exist', $object, $methodName), $exception);
 
@@ -265,16 +287,19 @@ class Exceptionist
      *
      * If the object owns the `has()` method, it uses that method. Otherwise, it
      *  uses the `property_exists()` function.
+     * @template ExistingProperty of string|string[]
      * @param object $object The class name or an object of the class to test for
-     * @param string|array<string> $property Name of the property or an array of names
-     * @param string|null $message The failure message that will be appended to
-     *  the generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return string|array<string>
-     * @throws \Tools\Exception\PropertyNotExistsException|\Throwable
+     * @param ExistingProperty $property Name of the property or an array of names
+     * @param string $message The failure message that will be appended to the generated message
+     * @param E|class-string<E> $exception The exception class you want to set
+     * @return ExistingProperty
+     * @throws \Tools\Exception\PropertyNotExistsException|\Exception
      */
-    public static function objectPropertyExists(object $object, $property, ?string $message = '', $exception = PropertyNotExistsException::class)
+    public static function objectPropertyExists(object $object, $property, string $message = '', $exception = PropertyNotExistsException::class)
     {
+        if (func_num_args() > 3) {
+            deprecationWarning('The `$exception` parameter is deprecated and will be removed in a later release');
+        }
         foreach ((array)$property as $name) {
             $result = method_exists($object, 'has') ? $object->has($name) : property_exists($object, $name);
             self::isTrue($result, $message ?: sprintf('Property `%s::$%s` does not exist', get_class($object), $name), $exception);
@@ -285,20 +310,26 @@ class Exceptionist
 
     /**
      * Checks whether a value is `true`
-     * @param mixed $value The value you want to check
-     * @param string|null $message The failure message that will be appended to the
-     *  generated message
-     * @param \Throwable|class-string<\Throwable> $exception The exception class you want to set
-     * @return mixed
-     * @throws \Exception|\Throwable
+     * @template TrueValue
+     * @param TrueValue $value The value you want to check
+     * @param string $message The failure message that will be appended to the generated message
+     * @param \Exception|class-string<\Exception> $exception The exception class you want to set
+     * @return TrueValue
+     * @throws \Exception
+     * @noinspection PhpConditionAlreadyCheckedInspection
      */
-    public static function isTrue($value, ?string $message = '', $exception = ErrorException::class)
+    public static function isTrue($value, string $message = '', $exception = ErrorException::class)
     {
         if ($value) {
             return $value;
         }
-        if (!$exception instanceof Throwable && !is_string($exception)) {
-            trigger_error('`$exception` parameter must be an instance of `Throwable` or a string');
+
+        if (is_object($exception)) {
+            deprecationWarning('Using an already instantiated exception is deprecated. Use only the exception class');
+        }
+
+        if (!$exception instanceof Exception && !is_string($exception)) {
+            trigger_error('`$exception` parameter must be an instance of `Exception` or a class string');
         }
 
         if (!$message) {
@@ -316,7 +347,7 @@ class Exceptionist
             }
         }
 
-        if ($exception instanceof Throwable) {
+        if ($exception instanceof Exception) {
             throw $exception;
         }